fix(HTD-41): move components to integration page
