+++ src/app/page.tsx
@@ -15,11 +15,9 @@ import {
   LinkSimpleBreak,
   GithubLogo,
   CursorClick,
-  DockerLogo,
   ShieldCheck,
   ArrowsLeftRight,
   Gear,
-  Wrench,
 } from '@/components/elements/icon';
 import ShowCase from '@/components/elements/showcase';
 import { SquareIcon } from '@/components/elements/square-icon';
@@ -38,7 +36,6 @@ import { DownloadButton } from '@/components/modules/download-button';
 import { Dropdown } from '@/components/modules/dropdown';
 import type { DropdownOptionProps } from '@/components/modules/dropdown/dropdown.types';
 import { FluidCard } from '@/components/modules/fluid-card';
-import type { FluidCardProps } from '@/components/modules/fluid-card/fluid-card.types';
 import { HeadingBlock } from '@/components/modules/heading-block';
 import { Input } from '@/components/modules/input';
 import { IntegrationCard } from '@/components/modules/integration-card';
@@ -46,12 +43,7 @@ import { NavigationSidebarLinks } from '@/components/modules/navigation-sidebar-
 import { TableContent } from '@/components/modules/table-content';
 import { CTA } from '@/components/sections/cta';
 import { FeatureLine } from '@/components/sections/feature-line';
-import { IntegrationCTA } from '@/components/sections/integration/cta';
-import { IntegrationGrid } from '@/components/sections/integration/grid';
-import { IntegrationBento } from '@/components/sections/integration/single-page/bento';
-import { IntegrationSinglePageHero } from '@/components/sections/integration/single-page/hero';
 import { IntegrationSteps } from '@/components/sections/integration/single-page/steps';
-import { IntegrationTextImage } from '@/components/sections/integration/single-page/text-image';
 import { Statistics } from '@/components/sections/statistics';
 import { Testimonials } from '@/components/sections/testimonials';
 import BlogPostImage from '@/content/posts/analytics-map.png';
@@ -132,28 +124,6 @@ export default async function Home() {
     },
   ];
 
-  const integrationsCard = [
-    {
-      $showBadge: true,
-      icon: AndroidLogo,
-      title: 'Android',
-      text: 'Android integration for apps and mobile browsers, including automated setup, per-app interception, and system-level certificate injection for complete visibility into emulators and rooted devices.',
-      link: {
-        href: '/example-1',
-        target: '_blank',
-      },
-    },
-    {
-      icon: AndroidLogo,
-      title: 'Javascript',
-      text: 'Android integration for apps and mobile browsers, including automated setup, per-app interception, and system-level certificate injection for complete visibility into emulators and rooted devices.',
-      link: {
-        href: '/example-2',
-        target: '_blank',
-      },
-    },
-  ];
-
   const navigationContentSidebar = [
     {
       text: 'Payments processing',
@@ -207,34 +177,6 @@ export default async function Home() {
       'Run any Python script, tool or server from that terminal',
       "Instantly see, debug & rewrite all Python's HTTP traffic",
     ],
-  ];
-
-  const integrationBentoCards: FluidCardProps[] = [
-    {
-      $variant: 'default',
-      title: 'Get started instantly',
-      text: 'No messing around with certificate files and wifi settings. \n\n Click the button, accept permissions, start capturing traffic.',
-      icon: RocketLaunch,
-    },
-    {
-      $variant: 'default',
-      title: "Debug your Android device's HTTP requests",
-      text: 'Scan a QR code on the device to start setup, or remotely connect debuggable devices via ADB. \n\n Reconnect again later in one tap. \n\n Supports Android Lollipop and later (v5 / API level 21+)',
-      icon: Wrench,
-    },
-    {
-      $variant: 'default',
-      title: 'Capture & inspect encrypted HTTPS',
-      text: 'Immediately view HTTPS on any device from apps that trust user-installed certificates, like Chrome.\n\n Enable trust in your own app with one tiny manifest change.\n\n On emulators & rooted devices, easily intercept HTTPS from any app, with automatic injection of a system certificate authority.',
-      icon: SealCheck,
-    },
-    {
-      $variant: 'highlighted',
-      title: 'Want the full details?',
-      text: 'Take a look at the in-depth guide to Android HTTP debugging.',
-      buttonText: 'Learn more',
-      buttonHref: '/learn-more',
-    },
   ];
 
   return (
@@ -622,44 +564,8 @@ export default async function Home() {
           <SimpleFooter />
         </ShowCase>
       </Container>
-      <IntegrationSinglePageHero
-        title="Http Toolkit and Docker integrated"
-        text="Intercept & view all Docker HTTP(S) Mock endpoints or entire servers Rewrite, redirect, or inject errors."
-        icon={DockerLogo}
-        breadcrumbText="docker"
-      />
-      <IntegrationGrid integrations={[...integrationsCard, ...integrationsCard, ...integrationsCard]} />
       <Statistics title="Why `*HTTP Toolkit*`?" text="Numbers that speak for themselves:" />
       <Testimonials />
-      <IntegrationCTA
-        $variant="hero"
-        title="Automatic setup and powerful integration for all your tools"
-        text="With deep integrations into a huge range of popular components & tools, HTTP Toolkit lets you intercept mobile apps, whole devices, bash scripts, entire Docker containers and more, so you can see & modify anything in just one click."
-      />
-      <IntegrationCTA
-        $variant="cta"
-        title="Integrations"
-        text="Explore a vast array of integrations and applications designed to streamline your work, consolidate information, and enhance collaboration effortlessly. "
-        button={{
-          $small: true,
-          $variant: 'secondary',
-          children: 'See all integrations',
-        }}
-      />
-      <IntegrationTextImage
-        title="HTTP Toolkit is a beautiful & open-source toolfor debugging, testing and building with HTTP(S)on Windows, Linux & Mac."
-        subtitle="what is http toolkit?"
-        image={{
-          darkSrc: '/images/mockup-image.webp',
-          lightSrc: '/images/mockup-image-light.webp',
-          alt: 'Mockup image',
-        }}
-      />
-      <IntegrationBento
-        title="HTTP Toolkit is a beautiful & open-source toolfor debugging, testing and building with HTTP(S)on Windows, Linux & Mac."
-        subtitle="what is http toolkit?"
-        cards={integrationBentoCards}
-      />
     </>
   );
 }
